test: rename test classes for MailboxServer
