+++ store/src/test/java/com/zextras/mailbox/MailboxServerTest.java
@@ -1,53 +1,101 @@
-/*
- * SPDX-FileCopyrightText: 2025 Zextras <https://www.zextras.com>
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
+// SPDX-FileCopyrightText: 2024 Zextras <https://www.zextras.com>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
 
 package com.zextras.mailbox;
 
-import com.zextras.mailbox.SampleLocalMailbox.ServerSetup;
-import com.zextras.mailbox.util.SoapClient;
-import com.zimbra.soap.account.message.AuthRequest;
-import com.zimbra.soap.type.AccountSelector;
-import org.apache.http.HttpResponse;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import com.zextras.mailbox.MailboxServer.Builder;
+import com.zextras.mailbox.MailboxServer.InstantiationException;
+import com.zimbra.cs.account.Config;
+import org.eclipse.jetty.server.Connector;
 import org.eclipse.jetty.server.Server;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class MailboxServerTest {
+  private Config config;
+  private com.zimbra.cs.account.Server localServer;
 
-	private static Server mailboxServer;
-	private static final int USER_PORT = 8080;
+  @BeforeEach
+  void setUp() {
+    config = mock(Config.class);
+    localServer = mock(com.zimbra.cs.account.Server.class);
+    when(localServer.getHttpContextPathBasedThreadPoolBalancingFilterRules()).thenReturn(new String[]{});
+    when(config.getHttpNumThreads()).thenReturn(20);
+    when(config.getMailboxdSSLProtocols()).thenReturn(new String[] {"TLS1.2"});
+    when(config.getSSLExcludeCipherSuites()).thenReturn(new String[] {"^TLS_RSA_.*"});
+    when(config.getSSLIncludeCipherSuites()).thenReturn(new String[] {});
+    when(config.getHttpConnectorMaxIdleTimeMillis()).thenReturn(3000);
+  }
 
-	@BeforeAll
-	static void setUp() throws Exception {
-		final String mailboxHome = MailboxServerTest.class.getResource("/").getFile();
-		final String timezoneFile = MailboxServerTest.class.getResource("/timezones-test.ics")
-				.getFile();
-		mailboxServer = new ServerSetup(USER_PORT, mailboxHome, timezoneFile).create();
-		mailboxServer.start();
-	}
+  @Test
+  void shouldCreateJettyServer_WithFourConnectors() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
 
-	@AfterAll
-	static void tearDown() throws Exception {
-		mailboxServer.stop();
-	}
+    final Connector[] connectors = server.getConnectors();
 
-	private String getUserEndpoint() {
-		return "http://localhost:" + USER_PORT + "/service/soap";
-	}
-	@Test
-	void shouldAuthenticateStandardUser() throws Exception {
-		try(SoapClient soapClient = new SoapClient(getUserEndpoint())) {
-			final HttpResponse httpResponse =  soapClient.newRequest()
-					.setSoapBody(new AuthRequest(AccountSelector.fromName("test@test.com"), "password"))
-					.execute();
+    assertEquals(4, connectors.length);
+  }
 
-			Assertions.assertEquals(200, httpResponse.getStatusLine().getStatusCode());
-		}
-	}
+  @Test
+  void shouldCreateJettyServer_WithOneHandler() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
+
+    assertEquals(1, server.getHandlers().length);
+  }
+
+  @Test
+  void shouldCreateJettyServer_withUserConnector() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
+
+    final Connector[] connectors = server.getConnectors();
+
+    final Connector userConnector = connectors[0];
+    assertEquals("http/1.1", userConnector.getProtocols().get(0));
+    assertEquals(3000, userConnector.getIdleTimeout());
+  }
+
+  @Test
+  void shouldCreateJettyServer_withAdminConnector() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
+
+    final Connector[] connectors = server.getConnectors();
+
+    final Connector adminConnector = connectors[1];
+    assertEquals(2, adminConnector.getProtocols().size());
+    assertEquals("ssl", adminConnector.getProtocols().get(0));
+    assertEquals("http/1.1", adminConnector.getProtocols().get(1));
+    assertEquals(0, adminConnector.getIdleTimeout());
+  }
+
+  @Test
+  void shouldCreateJettyServer_withMTAConnector() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
+
+    final Connector[] connectors = server.getConnectors();
+
+    final Connector adminMtaConnector = connectors[2];
+    assertEquals(2, adminMtaConnector.getProtocols().size());
+    assertEquals("ssl", adminMtaConnector.getProtocols().get(0));
+    assertEquals("http/1.1", adminMtaConnector.getProtocols().get(1));
+    assertEquals(0, adminMtaConnector.getIdleTimeout());
+  }
+
+  @Test
+  void shouldCreateJettyServer_withExtensionsConnector() throws InstantiationException {
+    final Server server = new Builder(config, localServer).build();
+
+    final Connector[] connectors = server.getConnectors();
+
+    final Connector extensionConnector = connectors[3];
+    assertEquals(2, extensionConnector.getProtocols().size());
+    assertEquals("ssl", extensionConnector.getProtocols().get(0));
+    assertEquals("http/1.1", extensionConnector.getProtocols().get(1));
+    assertEquals(3000, extensionConnector.getIdleTimeout());
+  }
 
 }
